CARDI-731 Extract out `FileClassifier`
When ghciwatch fails to start the underlying GHCi session (because of a
bulid failure in a dependent package), it currently exits. We'd like it
to wait around and attempt to restart GHCi after files change.

The logic for determining which file changes are worth restarting for is
currently inside the `Ghci` struct, meaning we can't use it once the
GHCi session dies.

Let's extract it so we can use it in the `Ghci` manager.
