+++ src/ghci/parse/show_paths.rs
@@ -1,5 +1,4 @@
 use std::collections::HashSet;
-use std::path::Path;
 
 use camino::Utf8Path;
 use camino::Utf8PathBuf;
@@ -31,11 +30,6 @@ pub struct ShowPaths {
 }
 
 impl ShowPaths {
-    /// Make a path relative to the working directory of this session.
-    pub fn make_relative(&self, path: impl AsRef<Path>) -> miette::Result<NormalPath> {
-        NormalPath::new(path, &self.cwd)
-    }
-
     /// Convert a target (from `:show targets` output) to a module source path.
     pub fn target_to_path(&self, target: &str) -> miette::Result<LoadedModule> {
         let target_path = Utf8Path::new(target);
